+++ test/models/product.rb
@@ -1,0 +1,5 @@
+class Product < ActiveRecord::Base
+  positioned
+
+  default_scope { order(:position) }
+end